requirements.txt
fastapi
uvicorn

langchain
langchain-ollama
langchain-community
langchain-google-genai
langserve
langchain-chroma
google-generativeai
python-dotenv

chromadb
pypdf

streamlit
requests

sse-starlette
python-multipart